Extract interface for service
