+++ spi/identity-hub-spi/src/test/java/org/eclipse/dataspaceconnector/identityhub/credentials/VerifiableCredentialsJWTServiceTest.java
@@ -31,7 +31,7 @@ import static org.eclipse.dataspaceconnector.identityhub.junit.testfixtures.Veri
 public class VerifiableCredentialsJWTServiceTest {
 
     static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-    static final VerifiableCredentialsJWTService VERIFIABLE_CREDENTIALS_JWT_SERVICE = new VerifiableCredentialsJWTService(OBJECT_MAPPER);
+    static final VerifiableCredentialsJWTService VERIFIABLE_CREDENTIALS_JWT_SERVICE = new VerifiableCredentialsJWTServiceImpl(OBJECT_MAPPER);
     static final Faker FAKER = new Faker();
     static final String VERIFIABLE_CREDENTIALS_KEY = "vc";
     static final JWSHeader JWS_HEADER = new JWSHeader.Builder(JWSAlgorithm.ES256).build();
